@@ -18,30 +18,29 @@
     "@angular/platform-browser": "^17.1.2",
     "@angular/platform-browser-dynamic": "^17.1.2",
     "@angular/service-worker": "^17.1.2",
-    "@dimforge/rapier3d": "^0.11.2",
-    "@dimforge/rapier3d-compat": "^0.11.2",
-    "@dotglitch/ngx-common": "^1.1.22",
-    "@egjs/flicking": "^4.11.2",
+    "@dagrejs/dagre": "^1.1.3",
+    "@dotglitch/ngx-common": "^1.1.48",
+    "@egjs/flicking": "^4.11.3",
     "@egjs/flicking-plugins": "^4.7.1",
-    "@egjs/ngx-flicking": "^4.11.2",
-    "@excalidraw/excalidraw": "^0.17.3",
-    "@fontsource/fira-mono": "^5.0.8",
-    "@fontsource/fira-sans": "^5.0.18",
-    "@fontsource/material-symbols-outlined": "^5.0.22",
+    "@egjs/ngx-flicking": "^4.11.3",
+    "@excalidraw/excalidraw": "^0.17.6",
+    "@fontsource/fira-mono": "^5.0.13",
+    "@fontsource/fira-sans": "^5.0.20",
+    "@fontsource/material-symbols-outlined": "^5.0.34",
     "@types/three": "^0.159.0",
-    "axios": "^1.6.7",
+    "axios": "^1.7.2",
     "localforage": "^1.10.0",
-    "matrix-animation": "^1.0.11",
-    "monaco-editor": "^0.45.0",
+    "matrix-animation": "^1.0.12",
+    "monaco-editor": "^0.50.0",
     "ngx-stackedit": "^1.1.55",
     "path-browserify": "^1.0.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "reactflow": "^11.10.1",
-    "rxjs": "~7.5.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "reactflow": "^11.11.4",
+    "rxjs": "~7.8.1",
     "three": "^0.159.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.14.2"
+    "tslib": "^2.6.3",
+    "zone.js": "~0.14.7"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.1.2",
@@ -50,11 +49,12 @@
     "@types/node": "^12.11.1",
     "@types/react": "^18.2.45",
     "@types/react-dom": "^18.2.18",
-    "angular-split": "^15.0.0",
-    "ngx-scrollbar": "^12.0.0",
-    "tabulator-tables": "^5.5.0",
-    "ts-node": "^10.9.1",
+    "angular-split": "^17.2.0",
+    "gsap": "^3.12.5",
+    "ngx-scrollbar": "^15.1.0",
+    "tabulator-tables": "^6.2.1",
+    "ts-node": "^10.9.2",
     "typescript": "~5.2.2",
-    "ulidx": "^2.2.1"
+    "ulidx": "^2.3.0"
   }
 }
